+++ code/+sibridge/getCurrentImage.m
@@ -19,11 +19,10 @@ function imData=getCurrentImage
 % Rob Campbell - Jan 2020
 
 
-im=[];
+imData = [];
 
 hSI = sibridge.getSIobject;
 if isempty(hSI)
-	imData = [];
     return
 end
 
@@ -43,11 +42,9 @@ imData = {};
 for ii = 1:nChannels
     tChan = stripeDataBuffer{1}.roiData{1}.imageData{ii}{1};
     tChan = repmat(tChan,1,1,nAveFrames);
-    size(tChan)
     for jj = 1:nAveFrames
         tChan(:,:,jj) = stripeDataBuffer{jj}.roiData{1}.imageData{ii}{1};
-
     end
-    imData{ii} = mean(tChan,3);
+    imData{ii} = mean(tChan,3).';
 end