+++ code/+sibridge/silinker.m
@@ -202,7 +202,7 @@ classdef silinker < handle
 
         function numPMTs = numberOfAvailablePMTs(obj)
             % Return the number of PMTs with an connected DAQ line as an integer
-            numPMTs =  cellfun(@(x) ~isempty(x.hAOGain), obj.hSI.hPmts.hPMTs)
+            numPMTs =  cellfun(@(x) ~isempty(x.hAOGain), obj.hSI.hPmts.hPMTs);
         end % numberOfAvailableChannels
 
 
@@ -218,7 +218,7 @@ classdef silinker < handle
                 return
             end
 
-            if length(gain)==1
+            if isscalar(gain)
                 obj.hSI.hPmts.gains = repmat(gain,1,4);
             elseif length(obj.hSI.hPmts.gains) == length(gain)
                 obj.hSI.hPmts.gains = gain(:)';
@@ -318,6 +318,12 @@ classdef silinker < handle
                 return
             end
 
+            if isempty(obj.hSI.hBeams.hBeams{beamIndex}.powerFraction2PowerWattLut)
+                % If empty this is an uncalibrated state so we need to make
+                % the array before it is populated
+                obj.hSI.hBeams.hBeams{beamIndex}.powerFraction2PowerWattLut = [0,0;1,1];
+            end
+            
             obj.hSI.hBeams.hBeams{beamIndex}.powerFraction2PowerWattLut(:,2) = minMaxW;
 
         end % setBeamMinMaxPowerInW